+++ pkg/deploy/ingress.go
@@ -93,7 +93,7 @@ func getGatewayPort(url string) (gateway string, port int, err error) {
 	gateway = split[0]
 	port, err = strconv.Atoi(split[1])
 	if err != nil {
-		return "", 0, fmt.Errorf("error ao pegar a porta do hostname: %s", err.Error())
+		return "", 0, fmt.Errorf("error ao identificar a porta do hostname: %s", err.Error())
 	}
 
 	return